requirements.txt
openai==1.12.0
flask==2.2.3
pyperclip==1.8.2
markdown==3.4.3
beautifulsoup4==4.12.2
python-dotenv==1.0.0
requests==2.31.0 